Fix velerobackup tests for DeepEquals
Several tests were using DeepEquals (under testify's `assert.Equal`) to
compare a runtime.Object from the (fake) kube server to a
locally-constructed expected value. This is brittle, as fields such as
ResourceVersion cannot be relied on to have predictable values. With
this commit, we switch to a comparison that disregards ResourceVersion
and TypeMeta.