Update e2e-test.sh for ClusterPools (step 1)
We have to do this in stages: Update this script such that it works for
both existing and new prow job config; update prow config; clean up this
script as necessary. In this stage:
- Tolerate a missing AWS creds file if the proper environment variables
are set. (For now, we have to use those env vars to generate the creds
file; left a TODO to fix that up.)
- Tolerate a missing SSH pubkey file, generating a keypair in this case.
- Detect provisioning failures when they happen, rather than timing out
after `installed` fails to appear for 90m.
- Remove some unnecessary variable `export`ing. Just because it bothered
me, not because it has any effect.