hibernation: include machines in undesired state when hibernating/resuming
Currently when hibernation is in progress, the controller checks if all
machines for the cluster are in desired state (stopped when hibernating
and running when running). If all machines are not in that state, the
controller returns and retries after some timeout.
This means that there is no information from the controller about the
progress made and machines in mis-matched state.

With this change, when machines are in mis-matched state, we update the
Hibernating condition to include a message listing those machines.

This is achieved by updating the hibernation.Actuator to return
mis-matched state machines from Machine{Running,Stopped} contract.