Safely delete ClusterDeployments
We're deleting ClusterDeployments from two different controllers:
clusterpool and clusterdeployment. This means there's the possibility of
timing problems whereby e.g.
- controller 1 loads a CD and decides it needs to be deleted
- controller 2 loads the same CD and updates it such that it should no
longer be deleted
- controller 1, having already decided to delete, deletes

This commit adds a `SafeDelete` utility that asks the server to check
the `ResourceVersion` of the object being deleted, and fail with a
`Conflict` error if it's out of date. It then changes all deletions of
ClusterDeployments to use that new method.