ClusterClaim handles CD deletion better
- Previously the ClusterClaim controller would only reconcile for
a deleted ClusterDeployment if the CD was actually gone. Now it also
notices if the CD has a `deletionTimestamp` or the
`remove-claimed-cluster-from-pool` annotation.
- Previously the ClusterDeleted condition was initialized to `Unknown`,
and then eventually set to `True`, but was never `False`. Now we set it
to `False` as soon as we ascertain the CD isn't deleted/deleting (per
above new logic).