hibernation: support hibernating spot instance for AWS clusters
OpenShift supports running spot instances for compute. Therefore when
cluster is hibernated these instances must also be stopped and started
like the on-demand instances.

But due to the nature of the spot instances, these instances are not
guaranteed. So if these instances are started, these might not start if
the spot instance constraints are not met. Therefore aws supports
stopping and starting spot instances only when the spot instance request
are persistent (compared to one-time). see [1] for more info.

machine-api add support for spot instances in AWS detailed in [2].
Some of the things to note are,
- machine-api has decided that one spot instance in aws is tied to one
  Machine object in api.
- which means, the spot instances are created as "one-time" type. Once
  the backing instance is gone (either terminated by aws or user) the
  Machine object goes into Failed state and stays there as-is.
- For new Machine to take the place of the previous Machine, the
  machine-api deploys MachineHealthCheck object that deletes Machines
  that were "deleted by AWS" only. If user deletes/terminated the
  instance, the user must also remove the object from API.

To shut down a "one-time" spot instance, the only option is to terminate
the instance. So for running -> hibernation, aws actuator terminated
these instances instead of stopping them like it does for on-demand
instances.

Since the spot instances were terminated, for hibernation -> running
state change the aws actuator must somehow replace the previous
instances with new ones.
The only way is to delete the Machine objects for the previously
terminated spot instance, and letting the MachineSet replace those to
match the expected replica count.
The aws actuator now looks for these Machines that need to be replaced
by
- finding machines with "machine.openshift.io/interruptible-instance"
  label
- filtering out machines that are "not failed" given that their status
  was updated after hibernation was started
The actuator then deletes those machines by
- updating the machine objects to skip draning. since the instances are
  already known to be gone, skipping draining prevents machine-api
  getting stuck for pods to scheduled.
- deleting those machine objects.

[1]: https://aws.amazon.com/about-aws/whats-new/2020/01/amazon-ec2-spot-instances-stopped-started-similar-to-on-demand-instances/
[2]: https://github.com/openshift/enhancements/blob/master/enhancements/machine-api/spot-instances.md