claim: Consolidate ClusterDeleted into ClusterRunning
Previously, the ClusterRunning condition of ClusterClaim was being used
to store the state of the assigned ClusterDeployment, *except* when the
CD was deleted. We had a whole separate condition, ClusterDeleted, for
that. This was particularly confusing in the summary output for
ClusterClaim, which could show something like:

```
[efried@efried ~]$ oc get clusterclaim claim-gllgp
NAME          POOL                     PENDING          CLUSTERNAMESPACE               CLUSTERRUNNING   AGE
claim-gllgp   fake-ocp-4-7-amd64-aws   ClusterClaimed   fake-ocp-4-7-amd64-aws-m2kkp   Running          4m23s
```

...even when the ClusterDeployment was deleted.

The reasonable fix is to twiddling the ClusterRunning condition to:

```
  {
    "lastProbeTime": "2021-07-16T20:02:01Z",
    "lastTransitionTime": "2021-07-16T20:02:01Z",
    "message": "Assigned cluster has been deleted",
    "reason": "ClusterDeleted",
    "status": "False",
    "type": "ClusterRunning"
  }
```

...at the same time as we set the ClusterDeleted condition to True. But that
would make the ClusterDeleted condition redundant, so this commit just gets rid
of it instead.

HIVE-1592