Don't allow excess claims to create extra CDs
Previously, due to a math error, if the number of pending claims
exceeded the pool size, we would create the surplus number of CDs,
assign them, and then delete them. This commit fixes the math error: any
excess claim will remain Pending until capacity becomes available to
create a CD for it.

**NOTE:**
The above also applies to the edge case where the pool size is zero.
A claim against such a pool will remain Pending *until the pool is
edited to have a nonzero size*.

This differs from other scenarios when we're at capacity:
- If we've reached `MaxSize`, capacity becomes available when a *CD* is
deleted (which may be prompted by its *claim* being deleted).
- If we've reached `MaxConcurrent`, capacity becomes available when a
*CD* finishes provisioning or deprovisioning.

Due to this discrepancy, the argument could be made to treat a zero-size
pool specially and simply fail any claim against it. We can debate that
separately; this commit doesn't address it.

HIVE-1593