Claim: Add ClusterRunning=>ClusterDeleting condition
This is a UX improvement to reflect that a CD is deleting in the status
of its assigned Claim: We set the ClusterRunning condition to False with
Reason: ClusterDeleting.

Note that, although this triggers on both deletion timestamp and the
`remove-claimed-cluster-from-pool` annotation, we are unlikely to notice
in the latter case because the claim itself is deleted almost
immediately in that code path. However, the code as written accounts for
a future where that deletion may wait on some other finalizer.