hiveconfig: allow choosing dns record type for private link
Currently the private link controller creates an ALIAS record in the PHZ
created for customer's cluster. This record points enables the dns
resolution of customer clsuters k8s API to the VPC endpoint, therefore
allowing hive to transparently accessing the cluster over the private
link.

In some cases the ALIAS record cannot be used. For example, in GovCloud
environments the Route53 private hosted zones do not support ALIAS
record type and therefore we have to use A records.

CNAME records would have been more suitable replacement for ALIAS
records, but
CNAME records cannot be created at the apex of a DNS zone like ALIAS
records. see https://stackoverflow.com/a/20230670
The private link architecture uses a DNS zone like
`api.<clustername>.<clusterdomain>`
so that there is no conflicts with authority on DNS resolution. Since
CNAME
records are not supported on the apex, we use A records pointing the
cluster's
API DNS name to the IP addresses of the VPC endpoint directly. Since
these IP
addresses do not change as they are backed by Elastic Network Interfaces
in the
corresponding subnets of the VPC endpoint, this DNS record should remain
stable.

When in A record mode, the controller extracts the ip addresses of the
vpc endpoint attached elastic networking interfaces and uses those ips
to create A record for the cluster's api dns name in the private hosted
zone.

For simplicity and current desired use-case, the dns record type can be
configured globally for all clusters managed by privatelink controller.

```yaml
spec:
  awsPrivateLink:
    dnsRecordType: (Alias | ARecord)
```