ClusterPool:ClusterDeployment version & condition
This commit adds a mechanism for calculating a hash representing the
following fields in ClusterPool.Spec:
- Platform
- BaseDomain
- ImageSetRef
- InstallConfigSecretTemplateRef

When creating a new ClusterDeployment for the pool, this hash is stored
in a new annotation on the ClusterDeployment.

The clusterpool controller checks all unassigned (installing or ready)
ClusterDeployments and sets a new condition on the ClusterPool. This
condition is:
- "True" when all CDs are up to date with the pool (including the edge
case where there are no CDs);
- "False" when one or more CDs are stale;
- "Unknown" if the annotation is missing from one or more CDs.
- (Also "Unknown" if we haven't had a chance to figure it out yet.)

Of note:
- When "False" or "Unknown", the condition's Message contains a list of
the offending CDs (in the spirit of ClusterSync's "Failed" condition).
- "False" takes precedence over "Unknown" (including: we don't try to
combine the CD lists in the Message).

HIVE-1058