Add templates, scripts, targets for OLM-less deploy
New `make` targets:
- `build-app-sre-template`: Makes sure hack/app-sre/saas-template.yaml
is up to date with the current state of the repo. Added to the `update`
chain. More below.
- `verify-app-sre-template`: Added to the `verify` chain. Makes sure the
above has been run.

The `build-app-sre-template` target
- Syncs the resources in the new `kustomization.yaml` to make sure we're
picking up all the CRDs under config/crds. The kustomization makes sure
the template will contain variables, which will be substituted by
app-sre via `oc process`, for the URI of the container image in the
deployment.
- Uses that kustomization to create a temporary `saas-objects.yaml` to
contain all the objects we wish to deploy via app-sre.
- Runs the new `generate-saas-template.py` script, which melds the
`saas-objects.yaml` with the new `saas-template-stub.yaml` to
generate/overwrite the new `saas-template.yaml`.
- Removes the temporary `saas-objects.yaml` file.

This `saas-template.yaml` file is what app-sre uses for direct deploy.

HIVE-1588