Refactor e2e-pool-test.sh
In preparation for adding more meat to the test, this refactors the
e2e-pool-test.sh script in a few ways:
- Factor out imageset creation into a function (future work will create
more than one)
- Generalize cleanup to delete all pools in the namespace (future work
will create more than one)
- Rework the algorithm that waits for a pool to be ready. Previously
this first waited for the expected number of CDs to be created; then
waited for them to have `.spec.installed == "true"`. With this commit we
instead rely on the clusterpool controller hiding all of that under
`.status.ready`, and we just wait for that to equal `.status.size`.
- Use a query that filters by pool name to discover the name of the
pool's (one) CD (future work will create more than one pool).