ClusterPool: Delete broken (ProvisionStopped) clusters
With this commit, the clusterpool controller detects "broken"
ClusterDeployments as those whose ProvisionStopped condition is True*
and deletes them so they can be replaced and do not continue to consume
capacity in the pool.

When adding or deleting clusters to satisfy capacity requirements, we
prioritize as follows:
- If we're under capacity, add new clusters before deleting broken
clusters. This is in case we hit `maxConcurrent`: we would prefer to use
that quota to add clusters to the pool.
- If we're over capacity, delete broken clusters before deleting viable
(installing or assignable) clusters.

*In the future, we may expand this definition to include other
definitions of "broken".

HIVE-1615