hive-operator: support exporting metrics to cluster
Adds a configuration to HiveConfig so that the operator can export
metrics for the controllers to prometheus instance running on the
cluster.

The operator deploys,
- the service monitors in target namespace for controllers,
- rbac resources so the in-cluster prometheus can access the monitoring
  resources and endpoints/services in the target namespace

An alternative was to use OLM supported method [1] to expose the
metrics. Hive uses CSV deployment method and the prometheus manifests
are only supported using operator bundle method. So using the OLM method
is not usefule.

[1]: https://github.com/openshift/enhancements/blob/master/enhancements/olm/olm-managed-operator-metrics.md