ClusterPool RunningCount
Adds ClusterPool.Spec.RunningCount, the number of pool clusters to try
to keep active at any time.

Note that when ClusterPool.Spec.HibernateAfter is set, we count from
when it was claimed rather than when it was started so it doesn't
hibernate early. Adds a
ClusterDeployment.Spec.ClusterPoolref.ClaimedTimestamp to track this.

HIVE-1576