controller/dnszone: add condition to track dns api opt-in required failures
Adds a special purpose condition to DNSZone like the
AuthenticationFailure condition to provide details when the controller
requires DNS apis to be enabled for success.

Currently, only the AWS actuator sets this condition based on request
from HIVE-1469.

The clusterdeployment controller tracks the conditions from DNSZone, and
it now also tracks the new condition to update the DNSReady condition on
the clusterdeployment.