Work around clustersync float compare test flake
We frequently see flakes like:

```
--- FAIL: TestReconcileClusterSync_NewSyncSetApplied (0.01s)
    assertion_compare.go:323:
        	Error Trace:	clustersync_controller_test.go:161
        	            				clustersync_controller_test.go:840
        	Error:      	"3599.994122811" is not greater than or equal to "3599.998926313"
        	Test:       	TestReconcileClusterSync_NewSyncSetApplied
        	Messages:   	[requeue after too small]
```

This isn't because the controller is doing something wrong; it's because
adding and substracting floating point numbers is subject to small
errors due to internal representation.

We don't really care about these requeue intervals beyond about 0.1s
granularity; so this commit truncates them to 2 decimal places to avert
this flake.