bundle-gen: Discover ocm-* branches, fix alpha
With this update, you can specify any commit-ish to `--branch` and we
will go look for `ocm-X.Y` branches on that commit. If we find any,
we'll add them to the list of channels to update. However, we'll error
if there's more than one and you're not tracking master, because then we
wouldn't know what semver prefix to use.

This also fixes a bug whereby we would erroneously update the `alpha`
channel if you specified an ocm branch that was a direct descendant of
`master`.

To summarize, assume the following (contrived) tree:
```
*  efef5656 (ocm-2.6, master)
*  cdcd3434 (ocm-2.4, ocm-2.5)
| * abab1212 (ocm-1.10, ocm-2.3)
| * 0101fafa (ocm-2.2)
|/
*
|
...
```
| `--branch`                               | Version       | Channels         |
| ---------------------------------------- | ------------- | ---------------- |
| None, `efef5656`, `ocm-2.6`, or `master` | 1.2.#-efef565 | alpha, ocm-2.6   |
| `cdcd3434`, `ocm-2.4`, or `ocm-2.5`      | 1.2.#-cdcd343 | ocm-2.4, ocm-2.5 |
| `abab1212`, `ocm-1.10`, or `ocm-2.3`     | Error         | Error            |
| `0101fafa`  or `ocm-2.2`                 | 2.2.#-0101faf | ocm-2.2          |

HIVE-1604