ProvisionDelay metric shouldn't log unknown conds
Provision underway metric logged conditions even if they were of unknown
status. Since we are initializing all the conditions now, it doesn't
seem appropriate to do so anymore.

This commit ensures the provision underway metric is logged if and only
if there are conditions present on the cluster deployment in undesired
state.