ClusterPool: running clusters for excess claims
In #1434, we added the paradigm of keeping some number of ClusterPool
clusters in Running state, so they don't have to suffer Resuming time
when claimed. However, that work was narrowly focussed on conforming to
the ClusterPool.Spec.RunningCount.

This commit expands the concept to try to make sure we're always running
enough clusters to satisfy claims. In particular, when we are servicing
a number of claims in excess of the pool's capacity, we create that
excess number of clusters... but we previously didn't account for that
number when calculating how many should be Running.

Now we do.

To explain via a corner case: If your pool has a zero Size and you
create a claim, we will create a cluster and immediately assign it. But
because RunningCount is zero, prior to this commit, we would create the
cluster with `PowerState=Hibernating`, and then kick it over to
`Running` when we assigned the claim. Now we'll create it with
`PowerState=Running` in the first place.

HIVE-1651