ClusterPool.Spec.InstallAttemptsLimit
Add `InstallAttemptsLimit` to the ClusterPool spec. This corresponds to
the ClusterDeployment.Spec field of the same name, and is copied into
pool-generated CDs when specified.

Note that hitting `InstallAttemptsLimit` results in the CD getting the
`ProvisionStopped` status condition set to `True`, and ClusterPool
responds to that by deleting (and replacing, as appropriate) the CD.

HIVE-1667