Unexport metrics
138c8e0 / #1525 / HIVE-1581 added `HiveConfig.spec.exportMetrics` and
logic to the operator to enable exporting metrics when this setting is
`true`. This commit adds logic to undo same when it is changed from
`true` to `false` (or removed).

We also add a watch on Namespaces matching
HiveConfig.Spec.TargetNamespace so we can re-reconcile the label if it
is edited manually.

HIVE-1655