operator: Allow secret watch to switch namespaces
The hive-operator establishes a Lister for watching Secrets in the
configured target namespace at runtime because we don't know the target
namespace at setup time. To avoid doing the same work on every
reconcile, this flow is gated by a flag which we set once we've
established the Lister.

Previously, this flag was a simple boolean. This meant that when
HiveConfig's targetNamespace was changed, we would not reconfigure the
Lister, and would continue watching for (and only being able to see, via
the Lister) secrets in the *old* namespace.

With this commit we change the flag to indicate the namespace for which
we established the Lister. Thus, when HiveConfig's targetNamespace
changes, we build a new one for the new namespace.

Part of HIVE-1656