operator: Label targetNamespace
In preparation for supporting changing HiveConfig's targetNamespace,
this commit adds a label to that namespace (whether it already existed
or not). Subsequent commits will use that label to identify stale target
namespaces and clean them up.

Note that this allows us to remove an earlier workaround for
https://github.com/kubernetes/kubernetes/issues/105689.

Part of HIVE-1656