Add cd.status.powerState
Previously we reported the powerState of cluster deployment via the
reason of hibernating condition. Add status.powerState to replace that.
This is in preparation for adding cluster ready condition. It then
allows us to explicitly set the powerState that is reported to users
when they get the cluster deployment.

powerState is set everytime we set the hibernating condition, only
exception is when we initialize the condition.