operator: scrub old target namespaces
In (hopefully) the final step of being able to edit
HiveConfig.Spec.TargetNamespace, this commit adds code to discover
previous target namespaces and scrub them of all the resources we might
have created therein.

I tried to make this as future-friendly as possible by putting the
cleanup for $thing as close as possible to the creation of $thing.
Hopefully future-me will notice this pattern when adding a new $thing
and remember to include cleanup code for it.

Closes HIVE-1656