Document targetNamespace quirk
Editing HiveConfig.Spec.TargetNamespace works, but leaves the old
namespace behind, even if we created it and you never messed with it.
This is (at least partly) because k8s/OpenShift grab hold of it and put
stuff in it that we shouldn't be in the business of trying to delete.
This commit documents this behavior in the CRD field description.

HIVE-1656