HiveConfig.Spec.FailedProvisionConfig.RetryReasons
Add support for a list of strings in HiveConfig which correspond to
`installFailingReason`s from [additional-]install-log-regexes.

- If the list is present and the most recent ClusterProvision has failed
with a matching reason, we will retry (assuming not otherwise blocked,
e.g. by `InstallAttemptsLimit`).
- If the list is present and the reason is _not_ in the list, we will
fail immediately (`ProvisionStopped`) even if `InstallAttemptsLimit` is
not reached.
- If the list is absent, the previous behavior is observed: we will
ignore the failure reason and proceed or stop based solely on factors
such as `InstallAttemptsLimit`.

HIVE-1662