Static name for managed DNS ConfigMap
When managed DNS is enabled, we copy the relevant data from HiveConfig
to a ConfigMap in the target namespace. Previously this ConfigMap was
dynamically named `managed-domains-XXXXX` and we would find it via a
label. For consistency with other similar ConfigMaps used by Hive, this
commit switches it to use a static name, `hive-managed-domains`,
instead. For upgrade purposes, we have to keep the old lookup mechanism
in place until we can be sure we've flushed out all instances of the old
naming scheme.