operator: DRY computing hashes
The operator computes hashes (MD5 sums) of ConfigMaps and Secrets and
uses those hashes to annotate the Deployments/StatefulSet for its child
controllers such that, if any of those ConfigMaps/Secrets change, the
annotations change, and the controllers get redeployed to pick up the
changes.

The code to compute these hashes, simple though it was, was duplicated
in a number of places. Consolidate to a single generic function.