e2e: Fix disappearing logs
We're saving deployment logs as part of an exit trap to ensure we get
the latest, most complete picture on failure. However, the way we were
doing it previously would truncate the log files if the deployments were
gone -- which they would be if the script didn't exit early, because we
uninstall the operator as the last part of the test.

Fix the log saving logic to only overwrite the log files if the log
extraction command succeeds.

We were also trying to grab the hive-operator logs from the wrong
namespace. Fix.