Fix hiveutil create-cluster --ssh-private-key-file
Specifying an SSH private key file to `hiveutil create-cluster` causes
it to parlay the contents into a Secret. However, the code is chomping
whitespace, including the trailing newline, which `ssh-add` seems to
require or it complains the file format is invalid. This commit
explicitly adds the trailing newline. This causes the resulting yaml
block scalar to be introduced with `|` now rather than `|-`, and
`ssh-add` works.