Stop assuming containers[0]
As hive-operator loads and modifies the manifests for the other
controllers, it used to assume that they only had one container. This is
not future-proof: if we added another container, it would be easy to
miss that we're now modifying the wrong one. Guard against this by
instead looking up the container by name.

There should be no functional change.