Set auth condition when auth validation fails.
Our previous test returned a nil value for the error received from
validating credentials which made it seem like we would always set an
auth failure condition. We actually just logged the error and
returned. The auth condition will now be set and a reconcile error
will be returned after (and when) the condition has been set.