ClusterClaim deletion waits for CD, Role[Binding]
Previously deleting an assigned ClusterClaim would cause the
clusterclaim controller to _mark_ the associated ClusterDeployment,
Role, and RoleBinding for deletion, then remove the ClusterClaim's
finalizer so it could be garbage collected.

Per the associated Jira card, this could cause confusion if the claim
was "resurrected" -- i.e. one with the same name was created. The new
claim would take the place of the old one, making it look like it was
"assigned" the existing, deleting ClusterDeployment.

With this change, we wait to remove the ClusterClaim's finalizer until
all the associated resources are actually 404. This should make it
impossible to "resurrect" a ClusterClaim: you won't be able to create a
claim with that name until all the associated resources are gone; so the
new claim will truly behave like a new claim.

HIVE-1729