ClusterPool: Claim running clusters
With this commit, the cdCollection bucket called `assignable` now only
contains installed, non-deleted/deleting, *running* clusters. Since we
fulfill claims out of that bucket, the result is that, when your claim
is assigned a cluster, that cluster is guaranteed to be running.
Before this change, it might have been hibernating and we would initiate
resume when assigning; so you would need to wait before using it.

To make this work, we also add a new bucket to cdCollection called
`standby`. This contains the installed, non-deleted/deleting clusters
that are *not* running.

The related metrics are changed accordingly:
- `hive_clusterpool_clusterdeployments_assignable` now only counts
running clusters.
- `hive_clusterpool_clusterdeployments_running` is removed, as it would
have been redundant with the above.
- `hive_clusterpool_clusterdeployments_standby` is added.

HIVE-1679