Use cd.Status.PowerState for ClusterClaim Running condition
The ClusterDeployment Hibernating condition used to be the one-stop shop
for determining the target cluster's power state. In #1604 we changed
the semantics such that the Hibernating condition only reports things
about hibernation; a new Ready condition reports things about being
ready; and a new Status.PowerState field reports the cluster's power
state definitively.

The ClusterClaim controller uses the ClusterDeployment's power state to
set the ClusterClaim's ClusterRunning condition. It was previously using
the state of the ClusterDeployment's Hibernating condition to do this.
We missed changing it in #1604 to look at the Ready condition or
Status.PowerState instead. This commit does the latter.

HIVE-1734