ClusterClaim: ClusterRunning cond => ClusterState status
Since #1652, we only ever assign claims to clusters that are already
running, so we don't need the ClusterRunning condition anymore.

However, some of the other information for which that condition was
being used -- namely, indicating when the cluster is deleting or
nonexistent -- is still useful, though doesn't make sense to be
associated with a thing called "ClusterRunning". So we add a
Status.ClusterState which can have one of three values:
- `NoCluster`: Either we haven't fulfilled the claim yet, or the
associated cluster has been deleted.
- `Provisioned`: The claimed cluster exists and is not being deleted
(though it is not necessarily running).
- `Deleting`: The claimed cluster is in the process of being deleted
(usually meaning it is deprovisioning.)

HIVE-1743