ClusterPool: Secondary sort by ns/name for running
The clusterpool controller sorts unclaimed clusters when deciding which
to make Running to satisfy `runningCount` and/or pending claims.
Generally we want FIFO behavior, so the primary sort is by
`creationTimestamp`. However, since that has a granularity of 1s, it is
possible for multiple CDs to have the same value, leading to
nondeterministic sorting on each reconcile and potential "thrashing" of
running-ness.

With this commit, we add a secondary sort by CD namespace and name to
force the sort to be deterministic.

HIVE-1747