bundle-gen: refactor: split CSV and package gen
Resolve a TODO by invoking the steps to generate the CSV and the package
file sequentially from the main routine instead of embedding the latter
in the former. This lets us simplify the parameter lists and paves the
way for a forthcoming commit to support generating the first bundle in a
new channel.