bundle-gen: Support overriding hive repo to clone
Last week I needed to work around a temporary upstream problem by using
a different base image in the Dockerfile. However, since bundle-gen
clones the hive repo from GitHub, I would have needed to commit and
merge that Dockerfile change, which would not be appropriate.

With this commit, bundle-gen accepts an optional `--hive-repo` parameter
allowing you to clone from something other than the default
(github/openshift/hive). In this case I used it to clone from a local
directory in which I had made the temporary Dockerfile update.

This has the ancillary benefit of allowing you to save time by pointing
to a local clone (but be careful that it's up to date!).