dnsendpoint: not-yet-scraped: no error, no requeue
Second attempt to reduce thrashing on controller startup with lots of
clusters.

Since the scraper posts an event causing the DNSZone to be enqueued,
there's no need for this path to requeue (immediately, or ever). And
since we're just waiting for the scrape, it's not really an error
condition.

HIVE-1855